Finished Chapter 20

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const { validationResult } = require('express-validator');
 
+const fileHelper = require('../utils/file');
 const Product = require('../models/mongoDb/product');
 const errorUtils = require('../utils/error');
 
@@ -60,7 +61,8 @@ exports.getEditProduct = (req, res, next) => {
 
 exports.postEditProduct = (req, res, next) => {
     const productId = req.body.productId;
-    const { title, imageUrl, description, price } = req.body;
+    const { title, description, price } = req.body;
+    const image = req.file;
     const errors = validationResult(req);
 
     if (!errors.isEmpty()) {
@@ -82,7 +84,10 @@ exports.postEditProduct = (req, res, next) => {
             }
 
             product.title = title;
-            product.imageUrl = imageUrl;
+            if (image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = `\\${image.path}`;
+            }
             product.description = description;
             product.price = price;
 
@@ -97,7 +102,19 @@ exports.postEditProduct = (req, res, next) => {
 };
 
 exports.postAddProduct = (req, res, next) => {
-    const { title, imageUrl, description, price } = req.body;
+    const { title, description, price } = req.body;
+    const image = req.file;
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: { title, description, price },
+            errorMessage: 'Attached file is not a image',
+            validationErrors: [],
+        });
+    }
     const errors = validationResult(req);
 
     if (!errors.isEmpty()) {
@@ -106,11 +123,13 @@ exports.postAddProduct = (req, res, next) => {
             path: '/admin/add-product',
             editing: false,
             hasError: true,
-            product: { title, imageUrl, description, price },
+            product: { title, description, price },
             errorMessage: errors.array()[0].msg,
             validationErrors: errors.array(),
         });
     }
+
+    const imageUrl = `\\${image.path}`;
 
     const product = new Product({
         title,
@@ -133,7 +152,13 @@ exports.postAddProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const productId = req.body.productId;
-    Product.deleteOne({ _id: productId, userId: req.user._id })
+    Product.findById(productId)
+        .then(product => {
+            if (!product) return next(new Error('Product not found!'));
+
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: productId, userId: req.user._id });
+        })
         .then(() => {
             console.log('Deleted Product!');
             res.redirect('/admin/product-list');
